requirements.txt
click==8.1.3
colorama==0.4.5
flake8==4.0.1
Flask==2.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mccabe==0.6.1
mypy==0.971
mypy-extensions==0.4.3
Pillow==9.2.0
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
tomli==2.0.1
typing_extensions==4.3.0
Werkzeug==2.1.2
